Refactor: UserModelが太りそうなのでUserDataを分離
